refactor: move date picker utils
