+++ packages/components/src/DatePicker/index.ts
@@ -1,3 +1,3 @@
 export * from "./DatePicker"
-export * from "./enums"
+export * from "../utils/date-controls/enums"
 export * from "./types"
