+++ packages/components/src/FilterDatePicker/subcomponents/FilterDatePickerField/filterDatePickerFieldReducer.tsx
@@ -1,4 +1,4 @@
-import { isInvalidDate } from "@kaizen/date-picker"
+import { isInvalidDate } from "~utils/date-controls"
 
 type DateFieldActions =
   | {
