+++ packages/components/src/FilterDatePicker/utils/transformDateToInputValue.ts
@@ -1,4 +1,5 @@
-import { DisabledDays, formatDateAsText } from "@kaizen/date-picker"
+import { DisabledDays } from "~types/date-controls"
+import { formatDateAsText } from "~utils/date-controls"
 
 export const transformDateToInputValue = (
   date: Date | undefined,
