+++ packages/components/src/FilterDateRangePicker/subcomponents/FilterDateRangePickerField/filterDateRangePickerFieldReducer.tsx
@@ -1,4 +1,4 @@
-import { isInvalidDate } from "@kaizen/date-picker"
+import { isInvalidDate } from "~utils/date-controls"
 
 type DateRangeFieldActions =
   | {
