+++ packages/components/src/types/date-controls/index.ts
@@ -1,2 +1,3 @@
+export * from "../../utils/date-controls/enums"
 export * from "./DisabledDays"
 export * from "./ReactDayPicker"
