+++ packages/components/src/utils/date-controls/index.ts
@@ -1,3 +1,9 @@
 export * from "./calculateDisabledDays"
+export * from "./formatDateAsNumeral"
+export * from "./formatDateAsText"
+export * from "./getLocale"
 export * from "./isDisabledDate"
 export * from "./isInvalidDate"
+export * from "./parseDateAsTextOrNumeral"
+export * from "./parseDateFromNumeralFormatValue"
+export * from "./parseDateFromTextFormatValue"
